scrapy/car_shop/requirements.txt
psycopg2
scrapy
python-dateutil
django